Mini Project: Box Model and Horizontal Aligment with Display and Float CSS Properties
Box Model Concept

It was mentioned that there are two different types of elements:
	Inline
	Block
The main things that differenciates them is that inline elements only take the width of the element that are inside them and block elements take the width of the whole element

It's it strongly advice to read more on this concept and learn it.
- display: inline-block
- float

As an exercise it was ask to the students to use any of this options for horizontal alignment of the elements on the mini project
It is also strongly adviced that they read more about how this two specific css properties work. Specially the float property.